feat: add metadata for key frontend pages
