+++ src/app/taxonomy/page.tsx
@@ -1,9 +1,27 @@
+import type { Metadata } from 'next';
 import Link from 'next/link';
 import { WikiNavBar } from '@/components/WikiNavBar';
 
 const API_URL =
   process.env.NEXT_PUBLIC_REPORIUM_API_URL ??
   'https://reporium-api-573778300586.us-central1.run.app';
+
+export const metadata: Metadata = {
+  title: 'Taxonomy Explorer - 8 AI skill dimensions',
+  description:
+    'Explore Reporium taxonomy coverage across skill areas, industries, use cases, modalities, AI trends, deployment context, tags, and maturity level.',
+  openGraph: {
+    title: 'Taxonomy Explorer - 8 AI skill dimensions',
+    description:
+      'Explore Reporium taxonomy coverage across skill areas, industries, use cases, modalities, AI trends, deployment context, tags, and maturity level.',
+    url: 'https://www.reporium.com/taxonomy',
+  },
+  twitter: {
+    title: 'Taxonomy Explorer - 8 AI skill dimensions',
+    description:
+      'Explore Reporium taxonomy coverage across skill areas, industries, use cases, modalities, AI trends, deployment context, tags, and maturity level.',
+  },
+};
 
 // ---------------------------------------------------------------------------
 // Types
